pyproject.toml
[project]
name = "gadeu"
version = "0.0.0"
description = "..a decorative auth library for Tornado."
authors = [
    { name="Shaun Wilson", email="mrshaunwilson@msn.com" }
]
keywords = ["tornado", "authentication", "authorization", "apiKey", "bearerToken", "JWT"]
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
classifiers = [
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Intended Audience :: Developers"
]
dependencies = [
    "base58>=2.1.1",
    "cryptography",
    "jwcrypto>=1.5.6",
    "tornado>=6"
]

[project.optional-dependencies]
dev = [
    "build",
    "punit>=1.1.4",
    "twine",
    "urllib3-future>=2.12.920"
]

[project.urls]
Documentation = "https://gadeu.readthedocs.io/"
Homepage = "https://github.com/wilson0x4d/gadeu"
Repository = "https://github.com/wilson0x4d/gadeu.git"

[build-system]
requires = ["build", "setuptools", "twine"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
exclude = [
    "docs",
    "tests",
    "tests.*"
]
